修改了一些东西

+++ src/main/java/com/petrichor/sincerity/mapper/SysUserMapper.java
@@ -1,7 +1,7 @@
 package com.petrichor.sincerity.mapper;
 
-import com.petrichor.sincerity.model.SysUserList;
-import com.petrichor.sincerity.model.SysUserRole;
+import com.petrichor.sincerity.vo.SysUserList;
+import com.petrichor.sincerity.vo.SysUserRole;
 import com.petrichor.sincerity.entity.SysPermission;
 import com.petrichor.sincerity.entity.SysUser;
 import org.apache.ibatis.annotations.Param;
@@ -17,7 +17,7 @@ public interface SysUserMapper {
 
     List<SysUserList> getUserList(SysUserList sysUserList);
 
-    Long insertUser(SysUser sysUser);
+    int insertUser(SysUser sysUser);
 
     int deleteUser(@Param("id") Long id, @Param("updateBy") String updateBy);
